+++ app/jobs/tally/sweeper_job.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Tally
+  class SweeperJob < ApplicationJob
+
+    def perform
+      Tally::Sweeper.sweep!
+    end
+
+  end
+end